+++ client/src/App.tsx
@@ -92,10 +92,13 @@ function App() {
 export default App;
 
 // TODO:
-// many and all msg delete feature - M
+// add reply feature - M
+// many and all message delete feature - M
 // make group chat feature like making private room and adding user to that room and user will get req if they accept they will be added to that room - H
 // make friend request feature where user can send friend request to other user and add them to group without permission - H
-// make friend list feature where user can see their friend list and chat with them - 
+// add feature to add user to group with permission - H
+// add feature to remove user from group - E
+// add feature to leave group - E
 // add web rtc for voice and video call feature - H
-// add notification feature for grp name, friend request, group request, etc - M
+// add notification feature for grp name, friend request, group request, etc - E