+++ History.md
@@ -1,3 +1,9 @@
+# 1.2.2 / 2025-03-08
+
+*Added Search bar in chat container
+*Added MOre option in chat container
+*Fixed minor bugs in CLient and Server for stability
+
 # 1.2.1 / 2025-03-07
 
 *Added version in api- (currently at v1)